Move networking tests to separate directory
