+++ tests/network/README.md
@@ -1,0 +1,6 @@
+This directory contains network tests that make use of network
+namespaces and should not be mixed with non-network-namespace-aware
+tests.
+
+For more info, see containernetworking/cni#262, vishvananda/netns#17
+etc.
