+++ tests/network/go.test
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+go test -v .
